requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-groq
langchain-core
langchain-community
groq
sqlalchemy
pymysql
pandas
sqlparse
python-dotenv
pydantic
